Add Choreo Trajectory Code
Choreo takes in the dynamics of swerve to output a feasible trajectory that will minimize time using CasADi. It is worth looking into and modifying in the offseason over PathPlanner. A big reason for this is PathPlanner switching over to Pure Pursuit, which is highly inaccurate. We can also try modifying PathPlanner to revert back its changes when it announces its beta period.