I am so sorry Spencer. Thank you for everything
Today we came to the conclusion that the particle filter is unmaintainable due to a lack of member experience, leading to a product that will ultimately become tech debt in one year after the current seniors leave. I am so sorry Spencer, thank you for all your hard work. I wish we didn't have to do this.